Create practise1.java
find the cards by using A&B  , if 11 means ne means next 1 cards is not a big number means add 1points,and 12 means ne means next 2 cards is not a big number means add 2points ,if 13 means ne means next 3 cards is not a big number means add 1points,and 14 means ne means next 4 cards is not a big number means add 4points. (A|B) is both cases have get same values if the above conditon is occurs. the finally print the total points of A and B.......////mahaprabu.s///